migrate graph extension payload model to models package
